requirements/base.txt
Django==3.1.4
django-environ==0.4.5
psycopg2-binary==2.8.6
uvicorn==0.13.1
djangorestframework==3.12.2
strawberry-graphql==0.56.2
django-cors-headers==3.7.0